+++ wgpu-hal/src/vulkan/adapter.rs
@@ -2547,6 +2547,12 @@ impl super::Adapter {
 
         let drop_guard = crate::DropGuard::from_option(drop_callback);
 
+        let empty_descriptor_set_layout = unsafe {
+            raw_device
+                .create_descriptor_set_layout(&vk::DescriptorSetLayoutCreateInfo::default(), None)
+                .map_err(super::map_host_device_oom_err)?
+        };
+
         let shared = Arc::new(super::DeviceShared {
             raw: raw_device,
             queue_families_indices: queue_families_indices.to_vec(),
@@ -2576,6 +2582,7 @@ impl super::Adapter {
 
             texture_identity_factory: super::ResourceIdentityFactory::new(),
             texture_view_identity_factory: super::ResourceIdentityFactory::new(),
+            empty_descriptor_set_layout,
         });
 
         let mut queues = Vec::new();